requirements.txt
asgiref==3.8.1
Django==5.2.3
django-debug-toolbar==5.2.0
django-redis==5.4.0
djangorestframework==3.16.0
environs==14.2.0
gunicorn==23.0.0
marshmallow==4.0.0
packaging==25.0
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
